finish commit

+++ controllers/thumbinator.js
@@ -8,6 +8,7 @@ const ffprobePath = require('@ffprobe-installer/ffprobe').path;
 const ffmpeg = require('fluent-ffmpeg');
 const ffprobe = require('ffprobe');
 const ffprobeStatic = require('ffprobe-static');
+const thumb = require('node-thumbnail').thumb;
 
 ffmpeg.setFfmpegPath(ffmpegPath);
 ffmpeg.setFfprobePath(ffprobePath);
@@ -20,23 +21,45 @@ module.exports = function(data) {
     
     return new Promise(function(resolve, reject) {
 
-        ffmpeg(incomingFile)
-            .on('end', function() {
-                console.log('Screenshots taken');
-                let obj = {...data, thumb: filename + ".jpg"}
-                resolve(obj);
-            })
-            .on('error', function(err) {
-                console.log(err);
-                stat(incomingFile).then(() => fs.unlink(incomingFile))
-                reject("Screenshot error");
-            })
-            .screenshots({
-                timestamps: [ '50%' ],
-                filename: filename + ".jpg",
-                size: '600x?',
-                folder: targetFolder,
-            })
+        if(data.mimetype.indexOf("video") !== -1){
+
+            ffmpeg(incomingFile)
+                .on('end', function() {
+                    console.log('Screenshots taken');
+                    let obj = {...data, thumb: "Thumb-" + filename + ".jpg"}
+                    resolve(obj);
+                })
+                .on('error', function(err) {
+                    console.log(err);
+                    stat(incomingFile).then(() => fs.unlink(incomingFile))
+                    reject("Screenshot error");
+                })
+                .screenshots({
+                    timestamps: [ '50%' ],
+                    filename: "Thumb-" + filename + ".jpg",
+                    size: '600x?',
+                    folder: targetFolder,
+                })
+
+        }else if(data.mimetype.indexOf("image") !== -1){
+
+            thumb({
+                source: incomingFile, 
+                destination: targetFolder,
+                width: 450,
+                concurrency: 4,
+                prefix: 'Thumb-',
+                suffix: ''
+            }, (files, err, stdout, stderr) => {
+                if (err) return reject(err)
+                let obj = {...data, thumb: "Thumb-" + filename}
+                resolve(obj)
+            });
+
+        }else{
+            reject("Ошибка загрузки - неизвестный формат данных") 
+        }
+
 
     })
